requirements.txt
# Core dependencies for all sessions
# Install with: pip install -r requirements.txt

# Session 01 — no API key required
tiktoken>=0.7.0                # Tokenization (offline, no API needed)
sentence-transformers>=3.0.0   # Local embeddings (downloads model ~90MB on first use)
numpy>=1.26.0                  # Vector math

# Session 02+ — Anthropic API key required (or use Ollama locally)
anthropic>=0.40.0              # Claude API client

# Session 03 — RAG & Vector Databases
chromadb>=0.5.0                # Local vector database (no server needed)
ollama>=0.3.0                  # Local LLM client (requires Ollama running: https://ollama.com)

# Utilities
python-dotenv>=1.0.0           # Load ANTHROPIC_API_KEY from .env file
